+++ pkg/grpc/util.go
@@ -13,7 +13,12 @@ limitations under the License.
 
 package grpc
 
-import commonv1pb "github.com/dapr/dapr/pkg/proto/common/v1"
+import (
+	"fmt"
+	"strings"
+
+	commonv1pb "github.com/dapr/dapr/pkg/proto/common/v1"
+)
 
 func stateConsistencyToString(c commonv1pb.StateOptions_StateConsistency) string {
 	switch c {
@@ -35,4 +40,26 @@ func stateConcurrencyToString(c commonv1pb.StateOptions_StateConcurrency) string
 	}
 
 	return ""
+}
+
+func getConfigSubscribeUniqueKey(storeName string, keys []string) string {
+	return fmt.Sprintf("%s||%s", storeName, strings.Join(keys, ","))
+}
+
+func getSubscribingKeys(storeName string, key string) []string {
+	storeNameAndKeys := strings.Split(key, "||")
+	if storeNameAndKeys[0] != storeName {
+		return []string{}
+	}
+	return strings.Split(storeNameAndKeys[1], ",")
+}
+
+func keyInKeysAndRemove(key string, keys []string) ([]string, bool) {
+	resultKeys := make([]string, 0)
+	for _, v := range keys {
+		if v != key {
+			resultKeys = append(resultKeys, v)
+		}
+	}
+	return resultKeys, len(resultKeys) != len(keys)
 }
